requirements.txt
schedule==0.5.0
requests==2.9.1
dhooks==1.1.0
discord.py==1.2.3
